重构User表-DAO-Service; 重构index页面(解耦合); 完善推文管理逻辑

+++ src/main/java/com/blog/controller/TweetController.java
@@ -22,6 +22,7 @@ import java.util.List;
 @RequestMapping("/tweet")
 public class TweetController {
 
+    public final String REJECT = "/error/401";
     @Autowired
     private TweetService tweetService;
 
@@ -40,20 +41,22 @@ public class TweetController {
     @Autowired
     private TweetFrontEndConvector tweetFrontEndConvector;
 
-    public void setTypeAndTag(Model model) {
-        model.addAttribute("types", typeService.getAllType());
+    public void setTag(Model model) {
         model.addAttribute("tags", tagService.getAllTag());
     }
 
-    @GetMapping("/list")  //后台显示博客列表
-    public String showTweetList(@RequestParam(required = false,defaultValue = "1",value = "pagenum")int pagenum, Model model){
-        PageHelper.startPage(pagenum, 5);
-        List<Tweet> tweets = tweetService.getAllTweet();
+    @GetMapping("/list")  //显示列表
+    public String showTweetList(@RequestParam(required = false,defaultValue = "1",value = "pagenum")int pagenum, Model model, HttpSession session){
+        User user = (User) session.getAttribute("user");
+        if(user == null){
+            return REJECT;
+        }
+        PageHelper.startPage(pagenum, 8);
+        List<Tweet> tweets = tweetService.getTweetsByUserId(user.getId());
         //得到分页结果对象
         PageInfo pageInfo = new PageInfo(tweets);
         model.addAttribute("pageInfo", pageInfo);
-        setTypeAndTag(model);  //查询类型和标签
-        //TODO : edit
+        setTag(model);  //查询类型和标签
         return "tweets";
     }
 
@@ -65,44 +68,48 @@ public class TweetController {
         PageInfo pageInfo = new PageInfo(Tweets);
         model.addAttribute("pageInfo", pageInfo);
         model.addAttribute("message", "查询成功");
-        setTypeAndTag(model);
-        //TODO : edit
+        setTag(model);
         return "tweets";
     }
 
     @GetMapping("/send") //去新增博客页面
-    public String toAddTweet(Model model){
+    public String toAddTweet(Model model, HttpSession session){
+        User user = (User) session.getAttribute("user");
+        if(user == null){
+            return "redirect:/login";
+        }
         model.addAttribute("tweet", new Tweet());  //返回一个tweet对象给前端th:object
-        setTypeAndTag(model);
+        setTag(model);
         return "send-tweet";
     }
 
     @GetMapping("/{id}/edit") //去编辑博客页面
-    public String toEditTweet(@PathVariable Long id, Model model){
+    public String toEditTweet(@PathVariable Long id, Model model, HttpSession session){
+        User user = (User) session.getAttribute("user");
         Tweet tweet = tweetService.getTweet(id);
-//        tweet.init();   //将tags集合转换为tagIds字符串
-        model.addAttribute("tweet", tweet);     //返回一个tweet对象给前端th:object
-        setTypeAndTag(model);
-        return "send-tweet";
+        if(user == null){
+            return "redirect:/login";
+        }
+        //如果不是该用户所发的推文，则跳转到错误页
+        if(tweet.getUserId().equals(user.getId())){
+            model.addAttribute("tweet", tweet);     //返回一个tweet对象给前端th:object
+            setTag(model);
+            return "send-tweet";
+        }
+        return REJECT;
     }
 
     @PostMapping("/") //新增、编辑博客
     public String addTweet(Tweet tweet, HttpSession session, RedirectAttributes attributes){
         //设置user属性
-//        tweet.setUser((User) session.getAttribute("user"));
         User user = (User) session.getAttribute("user"); //获取session中的user
+        if(user.getId()==null){
+            return "redirect:/login";
+        }
         //设置用户id
         tweet.setUserId(user.getId());
         //设置权限类别
-        tweet.setStatus(user.getStatus().toString());
-        //设置blog的type
-//        tweet.setTypeId(typeService.getType(tweet.getType().getId()));
-        //设置blog中typeId属性
-//        tweet.setTypeId(tweet.getType().getId());
-        //给blog中的List<Tag>赋值
-//        tweet.setTags(tagService.getTagByString(tweet.getTagIds()));
-        //TODO: Tag未完成
-
+        tweet.setStatus(user.getStatus());
         if (tweet.getId() == null) {   //id为空，则为新增
             tweetService.saveTweet(tweet);
             //getTagIds: 从前端传回的String类型，like 1,2,3
@@ -115,17 +122,20 @@ public class TweetController {
         }
 
         attributes.addFlashAttribute("msg", "新增成功");
-//        return "redirect:/admin/blogs";
         return "redirect:/";
     }
 
     @GetMapping("/{id}/delete")
-    public String deleteTweet(@PathVariable Long id, RedirectAttributes attributes){
-        tweetService.deleteTweet(id);
-        trendService.deleteByTweetId(id);
-        attributes.addFlashAttribute("msg", "删除成功");
-        return "redirect:/usercenter/tweets";
-//        return "redirect:/admin/blogs";
+    public String deleteTweet(@PathVariable Long id, RedirectAttributes attributes, HttpSession session){
+        User user = (User) session.getAttribute("user");
+        Tweet tweet = tweetService.getTweet(id);
+        if(tweet.getUserId().equals(user.getId())) {
+            tweetService.deleteTweet(id);
+            trendService.deleteByTweetId(id);
+            attributes.addFlashAttribute("msg", "删除成功");
+            return "redirect:/usercenter/tweets";
+        }
+        return REJECT;
     }
 
     @GetMapping("/likes/{id}")
@@ -154,7 +164,7 @@ public class TweetController {
             return "redirect:/login";
         }
         Tweet tweet = tweetService.getTweet(id);
-        //如果已经点赞，则取消点赞
+        //如果已经收藏，则取消收藏
         if(userCollectionService.isUserCollection(user.getId(), id)){
             userCollectionService.deleteUserCollection(user.getId(), id);
         }else {
